* libdb/db_lookup.c (make_content): Cast time_t value to long before   trying to print as %ld. * libdb/db_ndbm.c (ndbm_flclose, ndbm_flopen): Rename dbf to db to   avoid shadowing a global variable. * libdb/mydbm.h: Update prototypes and gdbm/ndbm macros likewise.

+++ libdb/db_lookup.c
@@ -239,7 +239,7 @@ datum make_content (struct mandata *in)
 		dash_if_unset (in->name),
 		in->ext,
 		in->sec,
-		in->_st_mtime,
+		(long) in->_st_mtime,
 		in->id,
 		in->pointer,
 		in->filter,
